hotel_bot/requirements.txt
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
requests
